refactor: break down the poa screen components
