+++ packages/wallets/src/features/accounts/screens/AddressSection/AddressSection.scss
@@ -1,0 +1,70 @@
+.wallets-address-section {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 1.6rem;
+    max-width: 85rem;
+
+    @include mobile {
+        width: 100%;
+    }
+
+    &__title {
+        display: flex;
+        height: 2.4rem;
+        justify-content: center;
+        align-items: center;
+        gap: 1.1rem;
+
+        @include mobile {
+            display: flex;
+            align-items: center;
+            gap: 0.8rem;
+            align-self: stretch;
+        }
+
+        &__divider {
+            width: 76rem;
+            height: 0.12rem;
+            flex-shrink: 0;
+            background: var(--system-light-7-secondary-background, #f2f3f4);
+
+            @include mobile {
+                height: 0.1rem;
+                flex: 1 0 0;
+            }
+        }
+    }
+    &__inline {
+        display: flex;
+        align-items: flex-start;
+        gap: 0.8rem;
+        margin-bottom: 1rem;
+
+        &-message {
+            width: 80rem;
+
+            @include mobile {
+                width: 100%;
+            }
+        }
+    }
+
+    &__input {
+        display: flex;
+        width: 84.2rem;
+        flex-direction: column;
+        gap: 1.6rem;
+
+        @include mobile {
+            width: 100%;
+        }
+
+        & .wallets-dropdown,
+        & .wallets-textfield {
+            @include desktop {
+                max-width: 84rem;
+            }
+        }
+    }
+}
