+++ packages/wallets/src/features/accounts/screens/AddressSection/index.ts
@@ -1,0 +1,1 @@
+export { default as AddressSection } from './AddressSection';
