+++ packages/wallets/src/features/accounts/screens/DocumentSubmission/index.ts
@@ -1,0 +1,1 @@
+export { default as DocumentSubmission } from './DocumentSubmission';
