+++ packages/wallets/src/features/accounts/screens/PoaScreen/PoaScreen.scss
@@ -1,0 +1,19 @@
+.wallets-poa {
+    width: 99.6rem;
+    height: 60rem;
+    display: flex;
+    padding: 1.6rem;
+    flex-direction: column;
+    align-items: center;
+    gap: 1.6rem;
+    flex: 1 0 0;
+    align-self: stretch;
+    background: var(--system-light-8-primary-background, #fff);
+    overflow-y: auto;
+
+    @include mobile {
+        width: 100%;
+        height: 100%;
+        padding: 1.6rem;
+    }
+}
