+++ packages/wallets/src/features/accounts/screens/PoaScreen/index.ts
@@ -1,0 +1,1 @@
+export { default as PoaScreen } from './PoaScreen';
