--- packages/wallets/src/features/accounts/screens/ResubmitPOA/ResubmitPOA.scss
@@ -1,209 +1,0 @@
-.wallets-poa {
-    width: 99.6rem;
-    height: 60rem;
-    display: flex;
-    padding: 1.6rem;
-    flex-direction: column;
-    align-items: center;
-    gap: 1.6rem;
-    flex: 1 0 0;
-    align-self: stretch;
-    background: var(--system-light-8-primary-background, #fff);
-    overflow-y: auto;
-
-    @include mobile {
-        width: 100%;
-        height: 100%;
-        padding: 1.6rem;
-    }
-
-    .wallets-textfield {
-        max-width: 84rem;
-
-        @include mobile {
-            max-width: 100%;
-        }
-    }
-
-    &__address {
-        display: flex;
-        flex-direction: column;
-        align-items: flex-start;
-        gap: 1.6rem;
-        max-width: 85rem;
-
-        @include mobile {
-            width: 100%;
-        }
-
-        &__title {
-            display: flex;
-            height: 2.4rem;
-            justify-content: center;
-            align-items: center;
-            gap: 1.1rem;
-
-            @include mobile {
-                display: flex;
-                align-items: center;
-                gap: 0.8rem;
-                align-self: stretch;
-            }
-
-            &__divider {
-                width: 76rem;
-                height: 0.12rem;
-                flex-shrink: 0;
-                background: var(--system-light-7-secondary-background, #f2f3f4);
-
-                @include mobile {
-                    height: 0.1rem;
-                    flex: 1 0 0;
-                }
-            }
-        }
-
-        &__inline {
-            display: flex;
-            align-items: flex-start;
-            gap: 0.8rem;
-
-            &-message {
-                width: 80rem;
-
-                @include mobile {
-                    width: 100%;
-                }
-            }
-        }
-
-        &__input {
-            display: flex;
-            width: 84.2rem;
-            flex-direction: column;
-            gap: 1.6rem;
-
-            @include mobile {
-                width: 100%;
-            }
-
-            & .wallets-dropdown,
-            & .wallets-textfield {
-                @include desktop {
-                    max-width: 84rem;
-                }
-            }
-        }
-    }
-
-    &__document {
-        display: flex;
-        flex-direction: column;
-        align-items: flex-start;
-        gap: 2.4rem;
-        max-width: 85rem;
-
-        @include mobile {
-            width: 100%;
-        }
-
-        &__title {
-            display: flex;
-            height: 2.4rem;
-            justify-content: center;
-            align-items: center;
-            gap: 1.1rem;
-
-            @include mobile {
-                display: flex;
-                align-items: center;
-                gap: 0.8rem;
-                align-self: stretch;
-            }
-
-            &__divider {
-                width: 66rem;
-                height: 0.12rem;
-                flex-shrink: 0;
-                background: var(--system-light-7-secondary-background, #f2f3f4);
-
-                @include mobile {
-                    height: 0.1rem;
-                    flex: 1 0 0;
-                }
-            }
-        }
-
-        &__container {
-            display: flex;
-            padding: 1.6rem 2.4rem;
-            flex-direction: column;
-            align-items: flex-start;
-            gap: 1.6rem;
-            align-self: stretch;
-            border-radius: 0.8rem;
-            border: 0.1rem solid var(--system-light-5-active-background, #d6dadb);
-
-            @include mobile {
-                border: none;
-                padding: 0;
-            }
-
-            &__disclaimer {
-                display: flex;
-                flex-direction: column;
-                align-items: flex-start;
-                gap: 1.6rem;
-                align-self: stretch;
-
-                &__list {
-                    list-style-type: disc;
-                    padding-left: 2rem;
-                }
-            }
-
-            &__upload {
-                display: flex;
-                flex-direction: column;
-                align-items: flex-start;
-                gap: 1.6rem;
-                align-self: stretch;
-
-                & > .wallets-dropzone__container {
-                    display: flex;
-                    flex-direction: column;
-                    justify-content: center;
-                    align-items: center;
-                    gap: 1.6rem;
-                    align-self: stretch;
-                    border-radius: 0.8rem;
-                    border: 0.1rem dashed var(--system-light-5-active-background, #d6dadb);
-                    cursor: pointer;
-                }
-
-                & .wallets-dropzone__placeholder {
-                    gap: 2rem;
-                }
-
-                & .wallets-dropzone__placeholder-icon {
-                    width: auto;
-                    height: auto;
-                }
-
-                &__requirements {
-                    display: flex;
-                    justify-content: space-between;
-                    align-items: flex-start;
-                    align-self: stretch;
-                }
-            }
-
-            &__common-mistakes {
-                display: flex;
-                flex-direction: column;
-                gap: 1.6rem;
-                width: 100%;
-            }
-        }
-    }
-}
