--- packages/wallets/src/features/accounts/screens/ResubmitPOA/index.ts
@@ -1,1 +1,0 @@
-export { default as ResubmitPOA } from './ResubmitPOA';
