+++ packages/wallets/src/features/accounts/screens/index.ts
@@ -1,6 +1,8 @@
+export * from './AddressSection';
 export * from './CommonMistakesExamples';
+export * from './DocumentSubmission';
 export * from './IDVDocumentUpload';
 export * from './ManualDocumentUpload';
 export * from './PersonalDetails';
-export * from './ResubmitPOA';
+export * from './PoaScreen';
 export * from './SelfieDocumentUpload';
